Use pluggy to hook into Xpublish
Works with  xarray-contrib/xpublish#146
