+++ requirements.txt
@@ -1,3 +1,3 @@
 opendap-protocol<1.2.0
 xarray
-xpublish<=0.1.0
+xpublish
