--- setup.cfg
@@ -1,50 +1,0 @@
-[metadata]
-name = xpublish_opendap
-description = My Awesome module
-author = AUTHOR NAME
-author_email = AUTHOR@EMAIL.COM
-url = https://github.com/ioos/ioos-python-package-skeleton
-long_description = file: README.md
-long_description_content_type = text/markdown
-license = BSD-3-Clause
-license_file = LICENSE.txt
-classifiers =
-    Development Status :: 5 - Production/Stable
-    Intended Audience :: Science/Research
-    Operating System :: OS Independent
-    License :: OSI Approved :: BSD License
-    Programming Language :: Python
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Topic :: Scientific/Engineering
-
-[options]
-zip_safe = False
-install_requires =
-    numpy
-    requests
-python_requires = >=3.6
-packages = find:
-
-[sdist]
-formats = gztar
-
-[check-manifest]
-ignore =
-    *.yml
-    *.yaml
-    .coveragerc
-    docs
-    docs/*
-    *.enc
-    notebooks
-    notebooks/*
-    tests
-    tests/*
-
-[flake8]
-max-line-length = 105
-ignore = E203, E501, W503
-exclude = xpublish_opendap/_version.py
