--- setup.py
@@ -1,12 +1,0 @@
-from setuptools import setup
-
-# CHANGE PKG NAME HERE
-pkg_name = "xpublish_opendap"
-
-setup(
-    use_scm_version={
-        "write_to": f"{pkg_name}/_version.py",
-        "write_to_template": '__version__ = "{version}"',
-        "tag_regex": r"^(?P<prefix>v)?(?P<version>[^\+]+)(?P<suffix>.*)?$",
-    },
-)
